ref: move settings to inits folder
