+++ organizer-backend-NET/Inits/AuthInit.cs
@@ -1,37 +1,13 @@
-﻿using Microsoft.EntityFrameworkCore;
-using organizer_backend_NET.DAL;
-using organizer_backend_NET.Implements.Services;
-using organizer_backend_NET.Implements.Interfaces;
-using organizer_backend_NET.DAL.Interfaces;
-using organizer_backend_NET.DAL.Repository;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
+﻿using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.IdentityModel.Tokens;
+using organizer_backend_NET.Settings;
 using System.Text;
 
-namespace organizer_backend_NET
+namespace organizer_backend_NET.Inits
 {
-    public static class Services
+    public static class AuthInit
     {
-        public static void InitRepositories(this WebApplicationBuilder builder)
-        {
-            builder.Services.AddScoped<ITodoRespository, TodoRepository>();
-            builder.Services.AddScoped<ICalendarRepository, CalendarRepository>();
-            builder.Services.AddScoped<IUserRepository, UserRepository>();
-        }
-
-        public static void InitServices(this WebApplicationBuilder builder)
-        {
-            builder.Services.AddScoped<ITodoService, TodoService>();
-            builder.Services.AddScoped<ICalendarService, CalendarService>();
-            builder.Services.AddScoped<IUserService, UserService>();
-        }
-
-        public static void InitDataBase(this WebApplicationBuilder builder)
-        {
-            builder.Services.AddDbContext<AppContextDb>(o => o.UseNpgsql(builder.Configuration.GetConnectionString("AppContextDb"))); //Integrated Security=true;Pooling=true;
-        }
-
-        public static void InitJWTAuth(this WebApplicationBuilder builder)
+        public static void Init(this WebApplicationBuilder builder)
         {
             builder.Services.Configure<JWTSettings>(builder.Configuration.GetSection("JWTSettings"));
