+++ organizer-backend-NET/Inits/RepositoriesInit.cs
@@ -1,0 +1,15 @@
+﻿using organizer_backend_NET.DAL.Interfaces;
+using organizer_backend_NET.DAL.Repository;
+
+namespace organizer_backend_NET.Services
+{
+    public static class RepositoriesInit
+    {
+        public static void Init(this WebApplicationBuilder builder)
+        {
+            builder.Services.AddScoped<ITodoRespository, TodoRepository>();
+            builder.Services.AddScoped<ICalendarRepository, CalendarRepository>();
+            builder.Services.AddScoped<IUserRepository, UserRepository>();
+        }
+    }
+}
